Move settings compat into separate namespace
